+++ include/objects/mesh.hpp
@@ -23,7 +23,7 @@ class Mesh : public Object {
   bool Intersect_Triangle(const Ray &ray, int tri, double &dist) const;
   void Read_Obj(const char *file);
   Box Bounding_Box(int part) const override;
-  void Update(float deltaT) override;
+  void Update(double deltaT) override;
 
 };
