+++ include/objects/object.hpp
@@ -54,7 +54,7 @@ class Object {
   // If part>=0, return the bounding box for the specified part.
   // If part<0, return the bounding box for the whole object.
   virtual Box Bounding_Box(int part) const = 0;
-  virtual void Update(float deltaT) = 0;
+  virtual void Update(double deltaT) = 0;
 
 };
