+++ include/objects/plane.hpp
@@ -17,7 +17,7 @@ class Plane : public Object {
   virtual Hit Intersection(const Ray &ray, int part) const override;
   virtual vec3 Normal(const vec3 &point, int part) const override;
   virtual Box Bounding_Box(int part) const override;
-  void Update(float deltaT) override;
+  void Update(double deltaT) override;
 
 };
