+++ include/objects/sphere.hpp
@@ -18,7 +18,7 @@ class Sphere : public Object {
   virtual Hit Intersection(const Ray &ray, int part) const override;
   virtual vec3 Normal(const vec3 &point, int part) const override;
   virtual Box Bounding_Box(int part) const override;
-  void Update(float deltaT) override;
+  void Update(double deltaT) override;
 
 };
