+++ src/objects/sphere.cpp
@@ -32,6 +32,6 @@ Box Sphere::Bounding_Box(int part) const {
   return box;
 }
 
-void Sphere::Update(float deltaT) {
+void Sphere::Update(double deltaT) {
 
 }
